Moved some stuff from RelaySelect to Select
On further consideration, the support for specifying which property to
use as key and label is useful outside of Relay as well, so this has
been moved to Select.

Because most items in this application have uuid as key, and name as
label, these are now the default props used.