Manually rewrote the sql structure
It's now far less bloated.
It no longer includes any auditing (was never inteneded to be included)
It uses column-level grants to prevent tampering with certain columns.
The updated_at timestamp is automatically updated by a trigger.
There is a role created (application) that has the minimal amount of
permissions required.